Unhelpful error message. Turns out I had swapped the key and the bucket name when trying to read the image bytes.
